Move SwerveSubsystem.java
subsystems/swervedrive/ was empty aside from this file so got rid of that folder
